Add dedicated leaderboard pages per sport
