+++ apps/web/src/app/leaderboard/[sport]/page.tsx
@@ -1,0 +1,5 @@
+import Leaderboard from "../leaderboard";
+
+export default function LeaderboardSportPage({ params }: { params: { sport: string } }) {
+  return <Leaderboard sport={params.sport} />;
+}
